backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.12
exceptiongroup==1.2.2
fastapi==0.115.8
fonttools==4.56.0
frozendict==2.4.6
graphql-core==3.2.6
h11==0.14.0
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
multitasking==0.0.11
numpy==2.0.2
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
strawberry-graphql==0.260.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
yfinance==0.2.54
zipp==3.21.0
